requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
comtypes==1.1.14
cryptography==38.0.4
Django==4.1.4
django-admin-berry==1.0.6
django-cors-headers==3.13.0
django-daraja==1.2.0
django-phonenumber-field==7.0.1
djangorestframework==3.14.0
Flask==2.2.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
MouseInfo==0.1.3
phonenumbers==8.13.3
Pillow==9.3.0
PyAudio==0.2.12
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
PyScreeze==0.1.28
python-decouple==3.6
python-dotenv==0.21.0
pyttsx3==2.90
pytweening==1.0.4
pytz==2022.7
pywhatkit==5.4
pywin32==305
reportlab==3.6.12
requests==2.28.1
soupsieve==2.3.2.post1
SpeechRecognition==3.9.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.13
Werkzeug==2.2.2
wikipedia==1.4.0
